work on selected12

+++ rendering.h
@@ -30,8 +30,8 @@
 #define LB 6 // left border - in pixels
 
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
-#define UCA (0.35 * MIN(CW, CH))
-#define UCB (0.75 * UCA)
+#define UCA ((int)(0.35 * MIN(CW, CH)))
+#define UCB ((int)(0.75 * UCA))
 #define UCAB (UCA - UCB)
 #define UCW 5
 #define UC_COLOR 0x07e0//under construction color
